requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
boto3==1.39.13
botocore==1.39.13
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
defusedxml==0.7.1
disnake==2.10.1
emoji==2.14.1
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
numpy==2.0.1
pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
tinycss2==1.4.0
urllib3==2.5.0
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
